Rename the pipelines, remove unnecessary code
